made updates

+++ app/views/apps/apps_add_all_news.py
@@ -2,6 +2,7 @@ from aiohttp import web
 import logging
 from config import Config
 import db
+from utils.check_permissions import is_permitted
 
 logger = logging.getLogger(__name__)
 config = Config()
@@ -9,6 +10,9 @@ config = Config()
 
 class AppsAddAllNewsView(web.View):
     async def get(self, *args, **kwargs):
+        user_permissions = is_permitted(self.request, ['apps_permission'])
+        if not user_permissions:
+            return web.HTTPMethodNotAllowed("", [])
         params = self.request.rel_url.query
 
         db.session.query(db.NewsAppsRelations).filter_by(app_id=params['id']).delete()
@@ -24,6 +28,10 @@ class AppsAddAllNewsView(web.View):
 
 class AppsDeleteAllNewsView(web.View):
     async def get(self, *args, **kwargs):
+        user_permissions = is_permitted(self.request, ['apps_permission'])
+        if not user_permissions:
+            return web.HTTPMethodNotAllowed("", [])
+
         params = self.request.rel_url.query
 
         db.session.query(db.NewsAppsRelations).filter_by(app_id=params['id']).delete()
